feat: duty cycle > 100 protection
Accomplises 2 things.

Firstly, it forces any values over 100 to be read as 100. This means
that even if the voltages grow out of tolerance or has more noise in a
future environment, it should not cause issues with fan instability.

Secondly, this adds a lag period so that even if a user sets it to
slightly less than 100, it will show 100 which is a state likely to be
set to frequently.